Implemented Role based access to the users

+++ express-server/controllers/memberController.ts
@@ -18,7 +18,11 @@ export const getMemberByIdHandler = async (
   if (isNaN(memberId)) {
     return response.status(400).json({ error: 'Invalid Member ID' });
   }
-
+  if (memberId !== request.id && request.role === 'user') {
+    return response
+      .status(403)
+      .json({ error: 'You are not allowed to access this information' });
+  }
   try {
     const member = await memberRepo.getById(memberId);
     if (member) {
@@ -39,7 +43,9 @@ export const listMembersHandler = async (
   const search = (request.query.search as string) || '';
   const limit = Number(request.query.limit) || 10;
   const offset = Number(request.query.offset) || 0;
-
+  if (request.role === 'user') {
+    return response.status(403).json({ error: 'Access denied' });
+  }
   try {
     const members = await memberRepo.list({
       search: search,
@@ -57,6 +63,9 @@ export const createMemberHandler = async (
   request: Request,
   response: Response
 ) => {
+  if (request.role === 'user') {
+    return response.status(403).json({ message: 'Access denied' });
+  }
   try {
     const memberData: IMemberBase = request.body;
     const result = await memberRepo.create(memberData);
@@ -75,6 +84,9 @@ export const updateMemberHandler = async (
   const memberId = Number(request.params.id);
   if (isNaN(memberId)) {
     return response.status(400).json({ error: 'Invalid Member ID' });
+  }
+  if (memberId !== request.id && request.role === 'user') {
+    return response.status(403).json({ error: 'Access denied' });
   }
 
   try {
@@ -99,7 +111,9 @@ export const deleteMemberHandler = async (
   if (isNaN(memberId)) {
     return response.status(400).json({ error: 'Invalid Member ID' });
   }
-
+  if (memberId !== request.id && request.role === 'user') {
+    return response.status(403).json({ error: 'Access denied' });
+  }
   try {
     const result = await memberRepo.delete(memberId);
     if (result) {
